[Enhancement] cache iceberg partition names when mv get updated partition names
