+++ fe/fe-core/src/main/java/com/starrocks/connector/PartitionUtil.java
@@ -61,6 +61,7 @@ import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hive.common.FileUtils;
 import org.apache.iceberg.PartitionField;
 import org.apache.iceberg.PartitionSpec;
+import org.apache.iceberg.Snapshot;
 import org.apache.iceberg.StructLike;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -254,8 +255,21 @@ public class PartitionUtil {
                 // return table name if table is unpartitioned
                 return Lists.newArrayList(icebergTable.getRemoteTableName());
             }
-            partitionNames = GlobalStateMgr.getCurrentState().getMetadataMgr().listPartitionNames(
-                    icebergTable.getCatalogName(), icebergTable.getRemoteDbName(), icebergTable.getRemoteTableName());
+
+            Optional<Snapshot> snapshot = icebergTable.getSnapshot();
+            // snapshot is null means the table is empty
+            if (!snapshot.isPresent()) {
+                return Lists.newArrayList();
+            }
+
+            if (snapshot.get().snapshotId() != icebergTable.getSnapshotId()) {
+                partitionNames = GlobalStateMgr.getCurrentState().getMetadataMgr().listPartitionNames(
+                        icebergTable.getCatalogName(), icebergTable.getRemoteDbName(), icebergTable.getRemoteTableName());
+                icebergTable.setCachedPartitionNames(partitionNames);
+                icebergTable.setSnapshotId(snapshot.get().snapshotId());
+            } else {
+                partitionNames = icebergTable.getCachedPartitionNames();
+            }
         } else if (table.isJDBCTable()) {
             JDBCTable jdbcTable = (JDBCTable) table;
             partitionNames = GlobalStateMgr.getCurrentState().getMetadataMgr().listPartitionNames(
